Adding support for Google Suite Directory Groups
Google Suite Directory Groups can be fetched for the user.
A white list of group emails can be defined grant access.
The list of groups can be passed to an NGINX variable which can be
passed to the applications for fine grained access control.
A service account with Google Suite Domain-Wide Delegation of Authority
is required to access Google Directory API.

Added the possibility to extract also user email and name from the Oauth
profile data and pass them on as NGINX variables.

Requires Lua module luaossl, which provides OpenSSL bindings.